Added OMS, LiftPositions and refactored structure

+++ src/main/java/frc/robot/subsystems/MedianFilter.java
@@ -4,34 +4,34 @@ import java.util.Arrays;
 
 public class MedianFilter 
 {
-    private float[] arrayForFilter;
+    private double[] arrayForFilter;
     private int filterPowerInit = 10;
     private int counter = 0;
-    private float prevResult = 0;
-    private float sum = 0;
+    private double prevResult = 0;
+    private double sum = 0;
 
     public MedianFilter()
     {
-        this.arrayForFilter = new float[this.filterPowerInit];
+        this.arrayForFilter = new double[this.filterPowerInit];
     }
 
     public MedianFilter(int filterPower)
     {
         this.filterPowerInit = filterPower;
-        this.arrayForFilter = new float[this.filterPowerInit];
+        this.arrayForFilter = new double[this.filterPowerInit];
     }
 
     public MedianFilter(int filterPower, int digit)
     {
         this.filterPowerInit = filterPower;
-        this.arrayForFilter = new float[this.filterPowerInit];
+        this.arrayForFilter = new double[this.filterPowerInit];
         for (int i = 0; i < this.arrayForFilter.length; ++i)
         { 
             this.arrayForFilter[i] = digit;
         }
     }
 
-    public float Filter(float val)
+    public double Filter(double val)
     {
         for (int i = filterPowerInit - 1; i > 0; i--)
         {
@@ -39,14 +39,14 @@ public class MedianFilter
         }
         this.arrayForFilter[0] = val;
 
-        float[] copiedArr = Arrays.copyOf(this.arrayForFilter, this.filterPowerInit);
+        double[] copiedArr = Arrays.copyOf(this.arrayForFilter, this.filterPowerInit);
         quickSort(copiedArr, 0, this.filterPowerInit - 1);
         return copiedArr[this.filterPowerInit / 2];
     }
 
-    public float adapFilter(float val){
-        float filVal = 0;
-        float k;
+    public double adapFilter(double val){
+        double filVal = 0;
+        double k;
         if(Math.abs(val - filVal) > 1.5){
             k = 0.9f;
         }else{
@@ -57,7 +57,7 @@ public class MedianFilter
         return filVal;
     }
 
-    public float midFilter(float val){
+    public double midFilter(double val){
         sum += val;
         counter++;
         if(counter == 10){
@@ -68,16 +68,16 @@ public class MedianFilter
         return prevResult;
     }
 
-    public float GetValue()
+    public double GetValue()
     {
-        float[] copiedArr = Arrays.copyOf(this.arrayForFilter, this.filterPowerInit);
+        double[] copiedArr = Arrays.copyOf(this.arrayForFilter, this.filterPowerInit);
         quickSort(copiedArr, 0, this.filterPowerInit - 1);
         return copiedArr[this.filterPowerInit / 2];
     }
 
     public void Reset()
     {
-        this.arrayForFilter = new float[this.filterPowerInit];
+        this.arrayForFilter = new double[this.filterPowerInit];
     }
 
     public void Reset(int digit)
@@ -88,7 +88,7 @@ public class MedianFilter
         }
     }
     
-    public void quickSort(float arr[], int begin, int end) {
+    public void quickSort(double arr[], int begin, int end) {
         if (begin < end) {
             int partitionIndex = partition(arr, begin, end);
      
@@ -97,21 +97,21 @@ public class MedianFilter
         }
     }
 
-    private int partition(float arr[], int begin, int end) {
-        float pivot = arr[end];
+    private int partition(double arr[], int begin, int end) {
+        double pivot = arr[end];
         int i = (begin-1);
      
         for (int j = begin; j < end; j++) {
             if (arr[j] <= pivot) {
                 i++;
      
-                float swapTemp = arr[i];
+                double swapTemp = arr[i];
                 arr[i] = arr[j];
                 arr[j] = swapTemp;
             }
         }
      
-        float swapTemp = arr[i+1];
+        double swapTemp = arr[i+1];
         arr[i+1] = arr[end];
         arr[end] = swapTemp;
